+++ crates/ruma-macros/src/identifiers/ruma_id/parse.rs
@@ -2,13 +2,13 @@
 
 use proc_macro2::Span;
 use quote::quote;
-use syn::parse_quote;
+use syn::{meta::ParseNestedMeta, parse_quote};
 
-use super::RumaId;
-use crate::{
-    identifiers::{RumaIdAttrs, SMALLVEC_INLINE_BYTES_DEFAULT, StorageCfgAttributes, Types},
-    util::RumaCommon,
-};
+use super::{RumaId, StorageCfgAttributes, Types};
+use crate::util::RumaCommon;
+
+/// The default size of the inline array for the `SmallVec` inner representation.
+const SMALLVEC_INLINE_BYTES_DEFAULT: usize = 32;
 
 impl RumaId {
     /// Parse the given `ruma_id` macro input.
@@ -73,5 +73,63 @@ impl RumaId {
             types,
             ruma_common,
         })
+    }
+}
+
+/// The parsed attributes of the `ruma_id` attribute macro.
+#[derive(Default)]
+pub(crate) struct RumaIdAttrs {
+    /// The path to the function to use to validate the identifier.
+    validate: Option<syn::Path>,
+
+    /// The size of the inline array for the `SmallVec` inner representation.
+    smallvec_inline_bytes: Option<usize>,
+}
+
+impl RumaIdAttrs {
+    /// Set the path to the function to use to validate the identifier.
+    ///
+    /// Returns an error if it is already set.
+    fn set_validate(&mut self, validate: syn::Path, meta: &ParseNestedMeta<'_>) -> syn::Result<()> {
+        if self.validate.is_some() {
+            return Err(meta.error("cannot have multiple values for `validate` attribute"));
+        }
+
+        self.validate = Some(validate);
+        Ok(())
+    }
+
+    /// Set the size of the inline array for the `SmallVec` inner representation.
+    ///
+    /// Returns an error if it is already set or if the value doesn't fit into a `usize`.
+    fn set_smallvec_inline_bytes(
+        &mut self,
+        inline_bytes: syn::LitInt,
+        meta: &ParseNestedMeta<'_>,
+    ) -> syn::Result<()> {
+        if self.smallvec_inline_bytes.is_some() {
+            return Err(
+                meta.error("cannot have multiple values for `smallvec_inline_bytes` attribute")
+            );
+        }
+
+        self.smallvec_inline_bytes = Some(inline_bytes.base10_parse()?);
+        Ok(())
+    }
+
+    /// Try to parse the given meta item and merge it into this `RumaIdAttrs`.
+    ///
+    /// Returns an error if an unknown `ruma_id` attribute is encountered, or if an attribute
+    /// that accepts a single value appears several times.
+    pub(crate) fn try_merge(&mut self, meta: ParseNestedMeta<'_>) -> syn::Result<()> {
+        if meta.path.is_ident("validate") {
+            return self.set_validate(meta.value()?.parse()?, &meta);
+        }
+
+        if meta.path.is_ident("smallvec_inline_bytes") {
+            return self.set_smallvec_inline_bytes(meta.value()?.parse()?, &meta);
+        }
+
+        Err(meta.error("unsupported `ruma_id` attribute"))
     }
 }