+++ crates/ruma-common/CHANGELOG.md
@@ -60,6 +60,16 @@ Breaking changes:
     returns the same type. The `user_id_ref!` macro behind the `unstable-identifier-ref-macros`
     cargo feature still allows to construct a `&'static UserId`.
   - `OwnedVoipId` was renamed to `VoipId`.
+- The `IdDst` derive macro was removed. The `ruma_id` attribute macro is available instead.
+  The difference between both macros is that `ruma_id` does not support different borrowed and
+  owned types, but implements a single owned type. The same attributes are supported for both
+  macros.
+
+  To transition from `IdDst` to `ruma_id` one only needs to:
+  - Remove the derive macros on the identifier struct.
+  - Remove `#[repr(transparent)]`.
+  - Leave or add the `ruma_id` attribute.
+  - Remove the struct fields declaration (e.g. `(str)`) after the struct name.
 
 Improvements: